requirements.txt
google-cloud-bigquery==3.11.4
google-cloud-aiplatform>=1.39.0
google-cloud-core==2.4.1
google-auth==2.27.0
google-auth-oauthlib==1.2.0
slack-bolt==1.18.0
python-dotenv==1.0.0
pandas==2.0.3
numpy==1.24.3
python-dateutil==2.8.2
typing-extensions==4.8.0
dspy-ai>=2.5.0
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
pyngrok==7.0.0
flask==2.3.3
requests==2.31.0 